Refactor ICD by subsystem Start to design coinacceptor endpoints
