+++ vmc/vmc-icd/src/lib.rs
@@ -3,63 +3,19 @@ use postcard_rpc::{endpoints, topics, TopicDirection};
 use postcard_schema::Schema;
 use serde::{Deserialize, Serialize};
 
+pub mod dispenser;
+use crate::dispenser::*;
+
+pub mod coinacceptor;
+use crate::coinacceptor::*;
+
+
 #[derive(Serialize, Deserialize, Schema, Debug, PartialEq)]
 pub struct ChillerStatus {
     temp_setpoint: f32,    
     current_temp: f32,
     chiller_on: bool,
     chiller_duty_cycle: u8,  //Averaged over preceding 24 hours as a %
-}
-
-
-#[derive(Serialize, Deserialize, Schema, Debug, PartialEq, Copy, Clone)]
-pub struct DispenserAddress {
-    pub row: char,
-    pub col: char,
-}
-
-//Information about a dispenser at a particular address
-//Will return None if the motor is not present
-pub type DispenserOption = Option<Dispenser>;
-#[derive(Serialize, Deserialize, Schema, Debug, PartialEq)]
-pub struct Dispenser {
-    pub address: DispenserAddress,
-    pub dispenser_type: DispenserType,
-    pub motor_status: MotorStatus,
-    pub can_status: Option<CanStatus>,
-}
-
-#[derive(Serialize, Deserialize, Schema, Debug, PartialEq)]
-pub enum DispenserType {
-    Spiral,
-    Can,
-}
-
-#[derive(Serialize, Deserialize, Schema, Debug, PartialEq)]
-pub enum MotorStatus {
-    Ok,
-    MotorNotHome,
-    //If motor not present, Dispenser Option would just be None
-}
-
-#[derive(Serialize, Deserialize, Schema, Debug, PartialEq)]
-pub enum CanStatus {
-    Ok,
-    LastCan,
-}
-
-//The result of attempting a vend operation
-pub type DispenseResult = Result<(), DispenseError>;
-
-#[derive(Serialize, Deserialize, Schema, Debug, PartialEq)]
-pub enum DispenseError {
-    MotorNotPresent,
-    MotorNotHome,
-    MotorStuckHome,
-    MotorStuckNotHome,
-    OneOrNoCansLeft, //Can vendor won't vend if only one can present
-    NoDropDetected,  //not implemented yet - my machine does not support
-    InvalidAddress,
 }
 
 endpoints! {
@@ -82,6 +38,8 @@ topics! {
 topics! {
     list = TOPICS_OUT_LIST;
     direction = TopicDirection::ToClient;
-    | TopicTy                   | MessageTy     | Path              | Cfg                           |
-    | -------                   | ---------     | ----              | ---                           |
+    | TopicTy                   | MessageTy     | Path                             | Cfg                           |
+    | -------                   | ---------     | ----                             | ---                           |
+    | CoinInsertedTopic         | CoinInserted  | "/mdb/coinacceptor/coininserted" |                               |
+    | EscrowPressedTopic        | ()            | "/mdb/coinacceptor/escrowpressed"|                               |
 }
